refactor: register.ts를 registration.ts로 파일명 변경
