+++ src/types/registration.ts
